User accee functionality added

+++ app/Http/Controllers/QuestionsController.php
@@ -28,8 +28,9 @@ class QuestionsController extends Controller
      */
     public function create()
     {
+
         $question = new Question();
-        return view('questions.create', compact($question));
+        return view('questions.create', compact('question'));
     }
 
     /**
@@ -67,6 +68,9 @@ class QuestionsController extends Controller
     {
         //$question = Question::findOrFail();
         //print_r($question);
+        if(\Gate::denies('update-question', $question)){
+            abort(403, 'Access denied');
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -92,6 +96,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, 'Access denied');
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted.');
